refactor(host): absorb session services
