+++ Cargo.toml
@@ -11,7 +11,6 @@ members = [
     "crates/core",
     "crates/engine",
     "crates/host",
-    "crates/session-services",
     "crates/llmsim",
     "crates/test-support",
     "crates/platform",
@@ -181,7 +180,6 @@ everruns-builtins = { path = "crates/builtins", version = "0.18.0" }
 everruns-core = { path = "crates/core", version = "0.18.0" }
 everruns-engine = { path = "crates/engine", version = "0.18.0" }
 everruns-host = { path = "crates/host", version = "0.18.0" }
-everruns-session-services = { path = "crates/session-services", version = "0.18.0" }
 everruns-llmsim = { path = "crates/llmsim", version = "0.18.0", default-features = false }
 everruns-test-support = { path = "crates/test-support", version = "0.18.0", default-features = false }
 everruns-platform = { path = "crates/platform", version = "0.18.0" }
