+++ crates/host/Cargo.toml
@@ -39,7 +39,6 @@ everruns-capability.workspace = true
 everruns-builtins = { workspace = true, optional = true }
 everruns-engine.workspace = true
 everruns-mcp = { workspace = true, optional = true }
-everruns-session-services.workspace = true
 everruns-provider.workspace = true
 everruns-integrations-filesystem = { workspace = true, optional = true }
 everruns-integrations-bashkit = { workspace = true, optional = true }
