+++ crates/host/src/session_services/capabilities/session_storage.rs
