+++ crates/host/src/session_services/session_mutator.rs
