+++ crates/platform/Cargo.toml
@@ -66,7 +66,6 @@ everruns-core = { workspace = true }
 everruns-host.workspace = true
 everruns-capability.workspace = true
 everruns-provider.workspace = true
-everruns-session-services.workspace = true
 everruns-builtins = { workspace = true, optional = true }
 everruns-integrations-filesystem = { workspace = true, optional = true }
 everruns-integrations-bashkit = { workspace = true, optional = true }
