+++ knowledge/framework/application-api.md
@@ -83,9 +83,9 @@ An ordinary application targets `everruns` alone for agent configuration and
 execution. The facade-only acceptance fixture enforces that boundary and is
 intentionally stricter than the requirement for a complete execution host.
 The default facade and `everruns-host --no-default-features` graphs do not
-contain `everruns-platform`, Reqwest, Rustls, or Hyper. The small shared
-`everruns-session-services` crate owns the session mutation/storage capability
-boundary below host and platform; hosted product services remain opt-in through
+contain `everruns-platform`, Reqwest, Rustls, or Hyper. `everruns-host` owns
+the session mutation/storage capability boundary and platform re-exports it;
+hosted product services remain opt-in through
 platform-enabled product composition.
 
 An advanced system integrator may combine `everruns` with `everruns-host` and
