+++ tests/fixtures/external-consumer/event-log/tests/spi.rs
@@ -23,7 +23,7 @@ use everruns_host::{
     InProcessRuntimeBuilder, RuntimeAgentStore, RuntimeHarnessStore, RuntimeProviderStore,
     RuntimeSessionStore,
 };
-use everruns_session_services::SessionMutator;
+use everruns_host::SessionMutator;
 use everruns_llmsim::{LlmSimConfig, LlmSimRuntimeExt};
 use external_event_log::ExternalEventLog;
